requirements.txt
langchain>=0.3
langchain-openai>=0.3
langchain-anthropic>=0.3
python-dotenv>=1.0
yfinance>=0.2
requests>=2.31
